fixed issues and improvements from first PR

+++ src/components/CRUDDash.tsx
@@ -1,11 +1,11 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import { FirebaseApp } from 'firebase/app'
 import { getFirestore, collection, addDoc, getDocs, deleteDoc, where, query, updateDoc } from 'firebase/firestore'
 import LoadingOverlay from './LoadingOverlay'
+import { SketchPicker } from 'react-color'
 
 interface CrudProps {
   app: FirebaseApp
-  isAdmin: boolean
 }
 
 const CRUDDash: React.FC<CrudProps> = ({ app }) => {
@@ -22,10 +22,10 @@ const CRUDDash: React.FC<CrudProps> = ({ app }) => {
   const [showModifyEntity, setShowModifyEntity] = useState(false)
   const [modifyEntityOldName, setModifyEntityOldName] = useState('')
   const [modifyEntityNewName, setModifyEntityNewName] = useState('')
-  const [modifyEntityNewSecondProperty, setModifyEntityNewSecondProperty] = useState('')
-  
+  const [modifyEntitySecondProperty, setModifyEntitySecondProperty] = useState('')
+
   const [errorMessage, setErrorMessage] = useState('')
-  const [entityType, setEntityType] = useState<'Categories' | 'Types'>('Categories') // Initialize with Category
+  const [entityType, setEntityType] = useState<'Categories' | 'Types'>('Categories')
 
   const showEntities = async () => {
     setIsLoading(true)
@@ -39,39 +39,60 @@ const CRUDDash: React.FC<CrudProps> = ({ app }) => {
       })
       setEntities(entitiesArray)
     } catch (error) {
-      setErrorMessage('Error fetching ' + entityType.toLowerCase())
+      setErrorMessage('Error al obtener ' + entityType.toLowerCase())
       window.alert(errorMessage)
     } finally {
       setIsLoading(false)
     }
   }
 
-  const handleEntityNameInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+  const handleAddEntityNameChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setAddEntityName(event.target.value)
   }
 
-  const handleAddEntitySecondPropertyInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setAddEntitySecondProperty(event.target.value)
+  const handleAddEntityColorChange = (color: any) => {
+    setAddEntitySecondProperty(color.hex) // Actualiza el valor del color
   }
 
- 
+  const handleModifyEntityNameChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setModifyEntityNewName(event.target.value)
+  }
+
+  const handleModifyEntityColorChange = (color: any) => {
+    setModifyEntitySecondProperty(color.hex)
+  }
+
   const handleAddEntity = async () => {
     try {
       if (addEntityName.trim() === '' || addEntitySecondProperty.trim() === '') {
-        // Do nothing if either input field is empty
-        return
+        return // No hagas nada si alguno de los campos de entrada está vacío
       }
-      if (entityType == 'Categories') {
-        await addDoc(collection(firestore, entityType), {
-          name: addEntityName,
-          color: addEntitySecondProperty,
-        })
-      } else {
-        await addDoc(collection(firestore, entityType), {
-          name: addEntityName,
-          categoryID: addEntitySecondProperty,
-        })
+
+      let secondProperty = addEntitySecondProperty
+
+      if (entityType === 'Types') {
+        // If entityType is 'Types', search for the category name in Firestore
+        const categoryQuery = query(collection(firestore, 'Categories'), where('name', '==', addEntitySecondProperty))
+        const categorySnapshot = await getDocs(categoryQuery)
+
+        if (!categorySnapshot.empty) {
+          // If category found, use its document ID as the category ID for 'Types'
+          categorySnapshot.forEach((doc) => {
+            secondProperty = doc.id
+          })
+        } else {
+          // Handle case when category not found
+          setErrorMessage('Categoría no encontrada')
+          window.alert(errorMessage)
+          return
+        }
       }
+
+      // Add the entity with the appropriate second property
+      await addDoc(collection(firestore, entityType), {
+        name: addEntityName,
+        categoryID: secondProperty,
+      })
 
       // Clear the input fields and hide them
       setAddEntityName('')
@@ -81,35 +102,34 @@ const CRUDDash: React.FC<CrudProps> = ({ app }) => {
       // Call showEntities to fetch and display the updated entities list
       await showEntities()
     } catch (error) {
-      setErrorMessage('Error adding ' + entityType.toLowerCase())
+      setErrorMessage('Error al agregar ' + entityType.toLowerCase())
       window.alert(errorMessage)
     }
   }
 
-  const toggleEntities = () => {
-    if (isEntitiesShown) {
-      setEntities([]) // Clear entities when hiding
-    } else {
-      showEntities() // Fetch entities when showing
-    }
-    setIsEntitiesShown(!isEntitiesShown) // Toggle the state
-  }
 
-  const handleDeleteEntity = async () => {
+
+  const handleDeleteEntity = async (entityToDelete: string) => {
     try {
-      const q = query(collection(firestore, entityType), where('name', '==', deleteEntityName))
-      const querySnapshot = await getDocs(q)
-      if (!querySnapshot.empty) {
-        querySnapshot.forEach(async (doc) => {
-          await deleteDoc(doc.ref)
-        })
-        await showEntities()
-      } else {
-        setErrorMessage(entityType + ' not found')
-        window.alert(errorMessage)
+      // Show confirmation dialog before deleting
+      const confirmed = window.confirm(`¿Estás seguro de que quieres eliminar ${entityToDelete}?`)
+
+      if (confirmed) {
+        // if confirmed, delete
+        const q = query(collection(firestore, entityType), where('name', '==', entityToDelete))
+        const querySnapshot = await getDocs(q)
+        if (!querySnapshot.empty) {
+          querySnapshot.forEach(async (doc) => {
+            await deleteDoc(doc.ref)
+          })
+          await showEntities()
+        } else {
+          setErrorMessage(entityType + ' no encontrado')
+          window.alert(errorMessage)
+        }
       }
     } catch (error) {
-      setErrorMessage('Error deleting ' + entityType.toLowerCase())
+      setErrorMessage('Error al eliminar ' + entityType.toLowerCase())
       window.alert(errorMessage)
     }
     setDeleteEntityName('')
@@ -118,156 +138,158 @@ const CRUDDash: React.FC<CrudProps> = ({ app }) => {
 
   const handleModifyEntity = async () => {
     if (modifyEntityNewName.trim() === '') {
-      setErrorMessage('Please enter a new name')
+      setErrorMessage('Por favor, ingresa un nuevo nombre')
       window.alert(errorMessage)
       return
     }
     try {
-      const q = query(collection(firestore, entityType), where('name', '==', modifyEntityOldName));
-      const querySnapshot = await getDocs(q);
+      const q = query(collection(firestore, entityType), where('name', '==', modifyEntityOldName))
+      const querySnapshot = await getDocs(q)
       if (!querySnapshot.empty) {
         querySnapshot.forEach(async (doc) => {
           if (entityType === 'Categories') {
             await updateDoc(doc.ref, {
               name: modifyEntityNewName,
-              color: modifyEntityNewSecondProperty,
-            });
+              color: modifyEntitySecondProperty,
+            })
           } else {
             await updateDoc(doc.ref, {
               name: modifyEntityNewName,
-              categoryID: modifyEntityNewSecondProperty,
-            });
+              categoryID: modifyEntitySecondProperty,
+            })
           }
-        });
+        })
       } else {
-        setErrorMessage(entityType + ' not found')
+        setErrorMessage(entityType + ' no encontrado')
         window.alert(errorMessage)
       }
-  
-      // Clear the input fields and hide them
-      setModifyEntityNewName('');
-      setModifyEntityNewSecondProperty('');
-      setModifyEntityOldName('');
-      setShowModifyEntity(false);
-  
-      // Call showEntities to fetch and display the updated entities list
-      await showEntities();
+
+      // clear and hide input fields
+      setModifyEntityNewName('')
+      setModifyEntitySecondProperty('')
+      setModifyEntityOldName('')
+      setShowModifyEntity(false)
+
+      // call showEntities to fetch and display updated entities
+      await showEntities()
     } catch (error) {
-      setErrorMessage('Error modifying ' + entityType.toLowerCase())
+      setErrorMessage('Error al modificar ' + entityType.toLowerCase())
       window.alert(errorMessage)
     }
-  };
-  
+  }
 
-  const toggleEntityType = (type: 'Category' | 'Types') => {
-    setEntityType(type)
-    setIsEntitiesShown(false) // Hide entities when switching type
-    showEntities() // Fetch entities for the new type
+  const toggleEntityType = async (type: 'Categories' | 'Types') => {
+    setEntityType(type) 
+    setIsEntitiesShown(false) 
+    setShowAddEntity(false)
+    setShowModifyEntity(false)
+  }
+
+  useEffect(() => {
+    showEntities() 
+  }, [entityType])
+
+  const entityTypesSpanish = {
+    Categories: 'Categorías',
+    Types: 'Tipos',
   }
 
   return (
     <div className="relative flex flex-col items-center justify-center h-full">
       <div className="absolute top-0 left-0 m-4">
-        <button onClick={() => toggleEntityType('Category')} className="px-4 py-2 mb-4 mr-2 text-white rounded-full shadow-md bg-green">
-          Switch to Category
+        <button onClick={() => toggleEntityType('Categories')} className="px-4 py-2 mb-4 mr-2 text-white rounded-full shadow-md bg-green">
+          Cambiar a Categoría
         </button>
         <button onClick={() => toggleEntityType('Types')} className="px-4 py-2 mb-4 mr-2 text-white rounded-full shadow-md bg-green">
-          Switch to Type
+          Cambiar a Tipo
         </button>
       </div>
+      {/* Title displaying current entity type */}
+      <h2 className="mb-4 text-2xl font-bold">{entityTypesSpanish[entityType]}</h2>
+      {/* Edit form */}
+      {showModifyEntity && (
+        <div className="flex flex-col items-center mb-4">
+          <input
+            type="text"
+            className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
+            placeholder={`Ingresa el nuevo nombre`}
+            value={modifyEntityNewName}
+            onChange={(e) => setModifyEntityNewName(e.target.value)}
+          />
+          {entityType === 'Categories' && <SketchPicker color={modifyEntitySecondProperty} onChange={handleModifyEntityColorChange} />}
+          {entityType !== 'Categories' && (
+            <input
+              type="text"
+              className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
+              placeholder="Ingresa el nuevo ID de categoría"
+              value={modifyEntitySecondProperty}
+              onChange={(e) => setModifyEntitySecondProperty(e.target.value)}
+            />
+          )}
+          <button onClick={handleModifyEntity} className="px-4 py-2 mr-2 text-white rounded-full shadow-md bg-green">
+            Guardar Cambios
+          </button>
+        </div>
+      )}
+      {/* Scrollable list view */}
+      <div className="w-full max-w-4xl mb-6 overflow-y-scroll ove6rflow-x-auto h-80">
+        {entities.map((entity, index) => (
+          <div key={index} className="flex items-center justify-between p-2 border-b">
+            <p>{entity}</p>
+            <div className="flex">
+              <button
+                onClick={() => {
+                  setModifyEntityOldName(entity)
+                  setShowModifyEntity(true)
+                  setShowAddEntity(false)
+                }}
+                className="px-4 py-2 ml-4 text-white rounded-full shadow-md bg-red"
+              >
+                Editar
+              </button>
+              <button onClick={() => handleDeleteEntity(entity)} className="px-4 py-2 ml-4 text-white rounded-full shadow-md bg-red">
+                Eliminar
+              </button>
+            </div>
+          </div>
+        ))}
+      </div>
+  
       <div className="flex flex-col items-center mb-4">
-        <button onClick={() => setShowAddEntity(true)} className="px-4 py-2 mb-4 text-white rounded-full shadow-md bg-red">
-          Add {entityType}
+        <button
+          onClick={() => {
+            setShowModifyEntity(false)
+            setShowAddEntity(true)
+          }}
+          className="px-4 py-2 mb-4 text-white rounded-full shadow-md bg-red"
+        >
+          Agregar {entityType}
         </button>
         {showAddEntity && (
           <div>
             <input
               type="text"
               className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
-              placeholder={`Enter ${entityType.toLowerCase()} name`}
+              placeholder={`Ingresa el nombre de ${entityType.toLowerCase()}`}
               value={addEntityName}
-              onChange={handleEntityNameInputChange}
+              onChange={handleAddEntityNameChange}
             />
-            <input
-              type="text"
-              className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
-              placeholder={entityType === 'Categories' ? 'Enter color code' : 'Enter category ID'}
-              value={addEntitySecondProperty}
-              onChange={handleAddEntitySecondPropertyInputChange}
-        
-            />
+            {entityType === 'Categories' && <SketchPicker color={addEntitySecondProperty} onChange={handleAddEntityColorChange} />}
+            {entityType !== 'Categories' && (
+              <input
+                type="text"
+                className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
+                placeholder="Ingresa el nombre de categoría"
+                value={addEntitySecondProperty}
+                onChange={(e) => setAddEntitySecondProperty(e.target.value)}
+              />
+            )}
             <button onClick={handleAddEntity} className="px-4 py-2 mr-2 text-white rounded-full shadow-md bg-green">
-              Submit
+              Enviar
             </button>
           </div>
         )}
       </div>
-      <div className="flex flex-col items-center mb-4">
-        <button onClick={() => setShowDeleteEntity(true)} className="px-4 py-2 mb-4 text-white rounded-full shadow-md bg-red">
-          Delete {entityType}
-        </button>
-        {showDeleteEntity && (
-          <div>
-            <input
-              type="text"
-              className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
-              placeholder={`Enter ${entityType.toLowerCase()} name`}
-              value={deleteEntityName}
-              onChange={(e) => setDeleteEntityName(e.target.value)}
-            />
-            <button onClick={handleDeleteEntity} className="px-4 py-2 mr-2 text-white rounded-full shadow-md bg-red">
-              Submit
-            </button>
-          </div>
-        )}
-      </div>
-      <div className="flex flex-col items-center mb-4">
-        <button onClick={() => setShowModifyEntity(true)} className="px-4 py-2 mb-4 text-white rounded-full shadow-md bg-red">
-          Modify {entityType}
-        </button>
-        {showModifyEntity && (
-          <div>
-            <input
-              type="text"
-              className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
-              placeholder={`old name`}
-              value={modifyEntityOldName}
-              onChange={(e) => setModifyEntityOldName(e.target.value)}
-            />
-            <input
-              type="text"
-              className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
-              placeholder={`new name`}
-              value={modifyEntityNewName}
-              onChange={(e) => setModifyEntityNewName(e.target.value)}
-       
-            />
-            <input
-              type="text"
-              className="px-3 py-2 mb-2 mr-2 border border-gray-300 rounded-md"
-              placeholder={entityType === 'Categories' ? 'new color code' : 'new category ID'}
-              value={modifyEntityNewSecondProperty}
-              onChange={(e) => setModifyEntityNewSecondProperty(e.target.value)}
-            />
-            <button onClick={handleModifyEntity} className="px-4 py-2 mr-2 text-white rounded-full shadow-md bg-green">
-              Submit
-            </button>
-          </div>
-        )}
-      </div>
-
-      <button onClick={toggleEntities} className="px-4 py-2 mb-4 text-white rounded-full shadow-md bg-red">
-        {isEntitiesShown ? `Hide ${entityType}` : `Show ${entityType}`}
-      </button>
-      {isEntitiesShown && (
-        <div className="grid w-full grid-cols-4 gap-3 p-4 text-center bg-red-100 rounded-lg max-w-8xl">
-          {entities.map((entity, index) => (
-            <p key={index} className="text-black">
-              {entity}
-            </p>
-          ))}
-        </div>
-      )}
       {isLoading && <LoadingOverlay isVisible={isLoading} />}
     </div>
   )
